Python/requirements.txt
flask
flask-restful
flask_login
Flask-SQLAlchemy
requests
mysql-connector
mysql-connector-python
pymysql
cryptography 